models/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0
pydantic>=1.10.0

# Phase 2 - ML Model Upgrades
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
prophet>=1.1.0
xgboost>=2.0.0
pymongo>=4.0.0
python-multipart>=0.0.6